+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Change Log
+- php-laravel: mobile index page view
+- ruby: new mobile index page view
+- python: new mobile index page view
+- php: new mobile index page view
+- java-spring: new mobile index page view
+- java: new mobile index page view
+- go: new mobile index page view
+- csharp-mvc: new mobile index page view
+- csharp: new mobile index page view
+- nodejs: new mobile index page view
 - update action icons
 - php-laravel: 5 editor icons on index page
 - java-spring: 5 editor icons on index page